Moved BruteForceResult
