+++ src/main/java/com/github/skjolberg/packing/BruteForcePackager.java
@@ -20,88 +20,6 @@ import java.util.List;
 
 public class BruteForcePackager extends Packager {
 
-	private static class BruteForceResult implements PackResult {
-
-		private PermutationRotationIterator rotator;
-		private List<Placement> items;
-		private Container container;
-
-		private int count;
-		private PermutationRotationState state;
-
-		public BruteForceResult(PermutationRotationIterator rotator, List<Placement> items, Container container) {
-			super();
-			this.rotator = rotator;
-			this.container = container;
-			this.items = items;
-		}
-
-		public boolean isRemainder() {
-			return count < items.size();
-		}
-
-		public Container getContainer() {
-			container.clear();
-			if(state == null ) {
-				throw new RuntimeException();
-			}
-			rotator.setState(state);
-
-			int result = pack(items, container, rotator, Long.MAX_VALUE, container, 0);
-			if(result == count) {
-				return container;
-			}
-			throw new IllegalArgumentException("Unexpected count " + result + ", expected " + count);
-		}
-
-		public void setState(PermutationRotationState state) {
-			this.state = state;
-		}
-
-		public void setCount(int count) {
-			this.count = count;
-		}
-
-		public int getCount() {
-			return count;
-		}
-
-		@Override
-		public boolean packsMoreBoxesThan(PackResult result) {
-			// return true if 'this' is better:
-			// - higher number of boxes
-			// - lower volume
-			// - lower max weight
-
-			BruteForceResult bruteForceResult = (BruteForceResult)result;
-			if(bruteForceResult.count < count) {
-				return true;
-			} else if(bruteForceResult.count == count) {
-				// check volume (of container)
-				if(bruteForceResult.container.getVolume() > container.getVolume()) {
-					return true;
-				} else if(bruteForceResult.container.getVolume() == container.getVolume()) {
-					// check weight (max weight of container, suboptimal but quick)
-					if(bruteForceResult.container.getWeight() > container.getWeight()) {
-						return true;
-					}
-				}
-			}
-
-			return false;
-		}
-
-		public PermutationRotationIterator getRotator() {
-			return rotator;
-		}
-
-		@Override
-		public boolean isEmpty() {
-			return count == 0;
-		}
-
-	}
-
 	/**
 	 * Constructor
 	 *
@@ -164,7 +82,7 @@ public class BruteForcePackager extends Packager {
 		return result;
 	}
 
-	protected static int pack(List<Placement> placements, Dimension container, PermutationRotationIterator rotator, long deadline, Container holder, int index) {
+	public static int pack(List<Placement> placements, Dimension container, PermutationRotationIterator rotator, long deadline, Container holder, int index) {
 	    if(placements.isEmpty()) {
 	    	return -1;
 	    }
@@ -218,7 +136,7 @@ public class BruteForcePackager extends Packager {
 		return true;
 	}
 
-	protected static int fit2D(PermutationRotationIterator rotator, int index, List<Placement> placements, Container holder, Placement usedSpace, long deadline) {
+	private static int fit2D(PermutationRotationIterator rotator, int index, List<Placement> placements, Container holder, Placement usedSpace, long deadline) {
 		// add used space box now
 		// there is up to 2 possible free spaces
 		holder.add(usedSpace);
@@ -277,7 +195,7 @@ public class BruteForcePackager extends Packager {
 		return fit2D(rotator, index, placements, holder, nextPlacement, deadline);
 	}
 
-	protected static boolean isFreespace(Space freespace, Box used, Placement target) {
+	private static boolean isFreespace(Space freespace, Box used, Placement target) {
 
 		// Two free spaces, on each rotation of the used space.
 		// Height is always the same, used box is assumed within free space height.
